COMP: Fix signed/unsigned conversion warning in vtkMRMLNodeTest1
svn-url: http://viewvc.slicer.org/viewvc.cgi/Slicer4?view=revision&revision=20666
git-svn-id: http://svn.slicer.org/Slicer4/trunk@20666 3bd1e089-480b-0410-8dfb-8563597acbee

+++ Libs/MRML/Core/Testing/vtkMRMLNodeTest1.cxx
@@ -56,7 +56,8 @@ int vtkMRMLNodeTest1(int , char * [] )
 //---------------------------------------------------------------------------
 bool TestSetAttribute(const char* attribute, const char* value,
                       const char* expectedValue,
-                      int expectedSize = 1, int expectedModified = 0)
+                      size_t expectedSize = 1,
+                      int expectedModified = 0)
 {
   vtkNew<vtkMRMLNodeTestHelper1> node;
   node->SetAttribute("Attribute0", "Value0");
